requirements.txt
streamlit==1.28.1
sqlite3-python==1.0.0
openai-whisper==20231117
PyPDF2==3.0.1
pydub==0.25.1
ffmpeg-python==0.2.1
requests==2.31.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
torch==2.0.1
tqdm==4.66.1
